requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=1.4.0
aiosqlite>=0.17.0
httpx>=0.24.0
python-multipart>=0.0.5
pydantic>=2.0.0
python-dotenv>=0.19.0
